# apps/docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
...
    "title": "Normal API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.normal.fast"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
    "/v1/api-keys": {
      "get": {
        "description": "List safe API Key metadata retained within the 90-day history window. Plaintext credentials are never redisplayed. Requires an active signed-in session and the exact web Origin.",
        "operationId": "listApiKeys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "api_keys": [
                    {
                      "connection_ids": [
                        "con_xxxxxxxxxxxxxxxxxxxxx"
                      ],
                      "created_at": "2026-08-14T12:00:00.000Z",
                      "credential_hint": "normal_apk_xxxxxxxxxxxxxxxxxxxxx.…wxyz",
                      "expires_at": null,
                      "id": "apk_xxxxxxxxxxxxxxxxxxxxx",
                      "last_used_at": null,
                      "name": "Nightly backup",
                      "permissions": [
                        "connections:read",
                        "directory:read"
                      ],
                      "revoked_at": null,
                      "state": "active"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyList"
                }
              }
            },
            "description": "Safe metadata for active, expired, and recently revoked API Keys."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": "not_found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "The signed-in session is missing or invalid."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "Identity or persistence authority is unavailable."
          }
        },
        "security": [
          {
            "clerkSession": []
          }
        ],
        "summary": "List API Key metadata",
        "tags": [
          "API Keys"
        ]
      },
      "post": {
        "description": "Create one named API Key for a signed-in User. Requires the exact web Origin, a Clerk bearer token, and first-factor verification within five minutes. The plaintext `normal_apk_` credential is returned only in this response. Grants are immutable after creation.",
        "operationId": "createApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "connection_ids": [
                  "con_xxxxxxxxxxxxxxxxxxxxx"
                ],
                "name": "Nightly backup",
                "permissions": [
                  "connections:read",
                  "directory:read"
                ]
              },
              "schema": {
                "$ref": "#/components/schemas/CreateApiKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "connection_ids": [
                    "con_xxxxxxxxxxxxxxxxxxxxx"
                  ],
                  "created_at": "2026-08-14T12:00:00.000Z",
                  "credential_hint": "normal_apk_xxxxxxxxxxxxxxxxxxxxx.…wxyz",
                  "expires_at": null,
                  "id": "apk_xxxxxxxxxxxxxxxxxxxxx",
                  "last_used_at": null,
                  "name": "Nightly backup",
                  "permissions": [
                    "connections:read",
                    "directory:read"
                  ],
                  "revoked_at": null,
                  "state": "active",
                  "credential": "normal_apk_<public-handle>.<secret-shown-once>"
                },
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKey"
                }
              }
            },
            "description": "The API Key was created. The `credential` field is the only plaintext copy."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "The body is invalid, the name is not unique among active keys, the selected Connections are not owned, or the active-key limit is reached."
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "clerk_error": {
                    "metadata": {
                      "reverification": {
                        "afterMinutes": 5,
                        "level": "first_factor"
                      }
                    },
                    "reason": "reverification-error",
                    "type": "forbidden"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ReverificationRequired"
                }
              }
            },
            "description": "First-factor verification is older than five minutes or missing."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "The signed-in session is missing or invalid."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "Identity or persistence authority is unavailable."
          }
        },
        "security": [
          {
            "clerkSession": []
          }
        ],
        "summary": "Create an API Key",
        "tags": [
          "API Keys"
        ]
      }
    },
    "/v1/api-keys/{api_key_id}": {
      "delete": {
        "description": "Permanently revoke one API Key owned by the signed-in User. Revocation is idempotent, clears the credential digest immediately, and takes effect on the next request. Requires an active signed-in session and the exact web Origin.",
        "operationId": "revokeApiKey",
        "parameters": [
          {
            "description": "Opaque `apk_` handle of the API Key to revoke.",
            "in": "path",
            "name": "api_key_id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^apk_[A-Za-z0-9_-]{21}$"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "api_key": {
                    "id": "apk_xxxxxxxxxxxxxxxxxxxxx",
                    "revoked_at": "2026-08-14T13:00:00.000Z",
                    "state": "revoked"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRevokeResponse"
                }
              }
            },
            "description": "The API Key is revoked and its digest is cleared. Repeating the call for an already revoked key owned by the User also succeeds."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "The signed-in session is missing or the handle is unknown to this User."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementError"
                }
              }
            },
            "description": "Identity or persistence authority is unavailable."
          }
        },
        "security": [
          {
            "clerkSession": []
          }
        ],
        "summary": "Revoke an API Key",
        "tags": [
          "API Keys"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiKeyList": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "required": [
          "api_keys"
        ],
        "properties": {
          "api_keys": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "connection_ids",
                "created_at",
                "credential_hint",
                "expires_at",
                "id",
                "last_used_at",
                "name",
                "permissions",
                "revoked_at",
                "state"
              ],
              "properties": {
                "connection_ids": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "description": "a string matching the pattern ^con_[A-Za-z0-9_-]{21}(?![\\s\\S])$",
                    "pattern": "^con_[A-Za-z0-9_-]{21}(?![\\s\\S])$"
                  }
                },
                "created_at": {
                  "type": "string",
                  "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                  "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                  "format": "date-time"
                },
                "credential_hint": {
                  "type": "string"
                },
                "expires_at": {
                  "anyOf": [
                    {
                      "type": "string",
                      "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                      "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                      "format": "date-time"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "id": {
                  "type": "string",
                  "description": "a string matching the pattern ^apk_[A-Za-z0-9_-]{21}(?![\\s\\S])$",
                  "pattern": "^apk_[A-Za-z0-9_-]{21}(?![\\s\\S])$"
                },
                "last_used_at": {
                  "anyOf": [
                    {
                      "type": "string",
                      "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                      "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                      "format": "date-time"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "name": {
                  "type": "string"
                },
                "permissions": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "connections:read",
                      "directory:read",
                      "messages:read",
                      "messages:send"
                    ]
                  }
                },
                "revoked_at": {
                  "anyOf": [
                    {
                      "type": "string",
                      "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                      "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                      "format": "date-time"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "state": {
                  "type": "string",
                  "enum": [
                    "active",
                    "expired",
                    "revoked"
                  ]
                }
              },
              "additionalProperties": false
            }
          }
        },
        "additionalProperties": false
      },
...
      "CreateApiKey": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "required": [
          "connection_ids",
          "name",
          "permissions"
        ],
        "properties": {
          "connection_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "a string matching the pattern ^con_[A-Za-z0-9_-]{21}(?![\\s\\S])$",
              "pattern": "^con_[A-Za-z0-9_-]{21}(?![\\s\\S])$"
            },
            "description": "an array of at most 50 item(s)",
            "title": "maxItems(50)",
            "minItems": 1,
            "maxItems": 50
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "connections:read",
                "directory:read",
                "messages:read",
                "messages:send"
              ]
            },
            "description": "an array of at most 4 item(s)",
            "title": "maxItems(4)",
            "minItems": 1,
            "maxItems": 4
          }
        },
        "additionalProperties": false
      },
...
      "CreatedApiKey": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "required": [
          "connection_ids",
          "created_at",
          "credential_hint",
          "expires_at",
          "id",
          "last_used_at",
          "name",
          "permissions",
          "revoked_at",
          "state",
          "credential"
        ],
        "properties": {
          "connection_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "a string matching the pattern ^con_[A-Za-z0-9_-]{21}(?![\\s\\S])$",
              "pattern": "^con_[A-Za-z0-9_-]{21}(?![\\s\\S])$"
            }
          },
          "created_at": {
            "type": "string",
            "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
            "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
            "format": "date-time"
          },
          "credential_hint": {
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "string",
            "description": "a string matching the pattern ^apk_[A-Za-z0-9_-]{21}(?![\\s\\S])$",
            "pattern": "^apk_[A-Za-z0-9_-]{21}(?![\\s\\S])$"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "connections:read",
                "directory:read",
                "messages:read",
                "messages:send"
              ]
            }
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "a string matching the pattern ^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?Z(?![\\s\\S])",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "active",
              "expired",
              "revoked"
            ]
          },
          "credential": {
            "type": "string",
            "description": "a string matching the pattern ^normal_apk_[A-Za-z0-9_-]{21}\\.[A-Za-z0-9_-]{43}(?![\\s\\S])$",
            "pattern": "^normal_apk_[A-Za-z0-9_-]{21}\\.[A-Za-z0-9_-]{43}(?![\\s\\S])$"
          }
        },
        "additionalProperties": false
      },
...
      "ManagementError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ReverificationRequired": {
        "additionalProperties": false,
        "properties": {
          "clerk_error": {
            "additionalProperties": false,
            "properties": {
              "metadata": {
                "additionalProperties": false,
                "properties": {
                  "reverification": {
                    "additionalProperties": false,
                    "properties": {
                      "afterMinutes": {
                        "type": "integer",
                        "const": 5
                      },
                      "level": {
                        "type": "string",
                        "const": "first_factor"
                      }
                    },
                    "required": [
                      "afterMinutes",
                      "level"
                    ],
                    "type": "object"
                  }
                },
                "required": [
                  "reverification"
                ],
                "type": "object"
              },
              "reason": {
                "type": "string",
                "const": "reverification-error"
              },
              "type": {
                "type": "string",
                "const": "forbidden"
              }
            },
            "required": [
              "metadata",
              "reason",
              "type"
            ],
            "type": "object"
          }
        },
        "required": [
          "clerk_error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "bearerFormat": "API Key",
        "description": "A `normal_apk_` split credential shown once at creation. Send it as `Authorization: Bearer <credential>` from a server. Never place it in a browser, query parameter, or cookie.",
        "scheme": "bearer",
        "type": "http"
      },
      "clerkSession": {
        "bearerFormat": "Clerk JWT",
        "description": "Signed-in product session for API Key management. Requires the exact configured web Origin and a Clerk bearer token. Creation also requires first-factor verification within five minutes. Server-side automations must not call these routes.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}